+++ snppipeline/snppipeline.py
@@ -321,7 +321,8 @@ def create_snp_matrix(options_dict):
         sample_name                  = os.path.basename(sample_directory)
         snp_pileup_file_path         = os.path.join(sample_directory, options_dict['pileupFileName'])
         position_consensus_base_dict = utils.create_consensus_dict(snp_pileup_file_path, options_dict['minConsFreq'])
-        verbose_print("Sample %s consensus base coverage = %.2f%%" % (sample_name, 100.0 * len(position_consensus_base_dict) / snplist_length))
+        if snplist_length > 0:
+            verbose_print("Sample %s consensus base coverage = %.2f%%" % (sample_name, 100.0 * len(position_consensus_base_dict) / snplist_length))
 
         snp_seq_string = ""
         for key in sorted_snp_position_list: